app/src/main/java/com/cannonballapps/notechaser/exercisesession/SessionViewModel2.kt
package com.cannonballapps.notechaser.exercisesession

import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import com.cannonballapps.notechaser.common.prefsstore.PrefsStore
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.StateFlow
import kotlinx.coroutines.flow.asStateFlow
import kotlinx.coroutines.flow.first
import kotlinx.coroutines.launch

class SessionViewModel2(
    prefsStore: PrefsStore,
) : ViewModel() {

    private val _sessionState = MutableStateFlow<SessionState>(SessionState.Loading)
    val sessionState: StateFlow<SessionState> = _sessionState.asStateFlow()

    init {
        viewModelScope.launch {
            prefsStore.playableGeneratorFlow().first()
            _sessionState.value = SessionState.Error
        }
    }
}

sealed interface SessionState {
    object Loading : SessionState
    object Error : SessionState
}
